% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toppFun.R
\name{get_Entrez}
\alias{get_Entrez}
\title{Convert genes into Entrez format}
\usage{
get_Entrez(genes)
}
\arguments{
\item{genes}{A list of genes}
}
\value{
a vector of genes in Entrez format
}
\description{
Convert genes into Entrez format
}
\examples{
get_Entrez(genes=c("IFNG", "FOXP3"))
}
