% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ifnb.markers.df}
\alias{ifnb.markers.df}
\title{IFNB Marker DF}
\format{
A dataframe with 100 rows and 14 columns
\describe{
\item{rank}{rank of marker}
\item{B}{cell group name}
\item{B Activated}{cell group name}
\item{CD14 Mono}{cell group name}
\item{CD16 Mono}{cell group name}
\item{CD4 Memory T}{cell group name}
\item{CD4 Naive T}{cell group name}
\item{CD8 T}{cell group name}
\item{DC}{cell group name}
\item{Eryth}{cell group name}
\item{Mk}{cell group name}
\item{CNK}{cell group name}
\item{pDC}{cell group name}
\item{T activated}{cell group name}
}
}
\source{
\url{https://www.nature.com/articles/nbt.4042}

Kang HM, Subramaniam M, Targ S, et al. Multiplexed droplet
single-cell RNA-sequencing using natural genetic variation.
Nat Biotechnol. 2018;36(1):89-94. doi:10.1038/nbt.4042
}
\usage{
data("ifnb.markers.df")
}
\description{
A dataframe of 100 top markers for each class in 'seurat_annotations' column
using presto::wilcoxauc() and presto::top_markers()
Created using the IFNB dataset from the SeuratData package
}
\keyword{datasets}
