% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ifnb.markers.list.CD8T}
\alias{ifnb.markers.list.CD8T}
\title{IFNB Marker DF}
\format{
A character vector with 100 genes
\describe{
\item{ifnb.markers.list.CD8T}{rank of marker}
}
}
\source{
\url{https://www.nature.com/articles/nbt.4042}
}
\usage{
data("ifnb.markers.list.CD8T")
}
\description{
A list of the 100 top markers for CD8 T cells in ifnb dataset
using presto::wilcoxauc() and presto::top_markers()
Created using the IFNB dataset from the SeuratData package
}
\keyword{datasets}
