% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scToppR-package.R
\docType{package}
\name{scToppR-package}
\alias{scToppR}
\alias{scToppR-package}
\title{scToppR: API Wrapper for ToppGene}
\description{
scToppR provides an easy-to-use API wrapper for the ToppGene web platform, used for gene ontology and functional enrichment research. The package also integrates visualization tools, making it a convenient tool directly connecting ToppGene to code-based workflows in R. The tool can also easily save results into different formats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BioinformaticsMUSC/scToppR}
  \item Report bugs at \url{https://github.com/BioinformaticsMUSC/scToppR}
}

}
\author{
\strong{Maintainer}: Bryan Granger \email{grangerb@musc.edu} (\href{https://orcid.org/0009-0008-6663-3755}{ORCID})

}
\keyword{internal}
