% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topp_plot.R
\name{toppBalloon}
\alias{toppBalloon}
\title{Create a balloon plot from toppdata results}
\usage{
toppBalloon(
  toppData,
  categories = NULL,
  balloons = 3,
  x_axis_text_size = 6,
  cluster_col = "Cluster",
  filename = "toppBalloon",
  save = FALSE,
  save_dir = tempdir(),
  height = 6,
  width = 8,
  slot_name = "toppData",
  ...
)
}
\arguments{
\item{toppData}{A toppData results dataframe, SummarizedExperiment, or SingleCellExperiment object}

\item{categories}{The topp categories to plot}

\item{balloons}{Number of balloons per group to plot}

\item{x_axis_text_size}{Size of the text on the x axis}

\item{cluster_col}{The column name for clusters (default: "Cluster")}

\item{filename}{Filename of the saved balloon plot}

\item{save}{Save the balloon plot if TRUE}

\item{save_dir}{Directory to save the balloon plot}

\item{height}{Height of the saved balloon plot}

\item{width}{Width of the saved balloon plot}

\item{slot_name}{For SE/SCE objects, the metadata slot name containing toppData (default: "toppData")}

\item{...}{Additional parameters for future use}
}
\value{
ggplot object or list of ggplot objects
}
\description{
This function creates balloon plots from ToppGene enrichment results. It accepts either
a data.frame with toppData results, or a SummarizedExperiment/SingleCellExperiment
object with toppData stored in the metadata.
}
\examples{
data("toppdata.pbmc")

# With data.frame
toppBalloon(toppdata.pbmc, balloons = 3, save = FALSE)

# With SummarizedExperiment (if toppData stored in metadata)
# toppBalloon(se_object, categories = "GeneOntologyMolecularFunction")

}
