% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toppdata.airway}
\alias{toppdata.airway}
\title{toppData example using the airway dataset results}
\format{
A dataframe with 902 rows and 14 columns
\describe{
\item{Category}{ToppGene category}
\item{ID}{ToppGene Term ID}
\item{Name}{ToppGene Term Name}
\item{PValue}{P value}
\item{QValueFDRBH}{adjusted p-value (FDR)}
\item{QValueFDRBY}{adjusted p-value (BY)}
\item{QValueBonferroni}{adjusted p-value (Bonferroni)}
\item{TotalGenes}{Total genes in background}
\item{GenesInTerm}{Genes in ToppGene Term}
\item{GenesInQuery}{Genes in submitted query}
\item{GenesInTermQuery}{Intersection of genes in Term and in Query}
\item{Source}{ToppGene result source}
\item{URL}{ToppGene associated URL}
\item{Cluster}{cell group name}
}
}
\source{
\url{https://toppgene.cchmc.org}

Generated using ToppGene API (\url{https://toppgene.cchmc.org/}).
Chen J, Bardes EE, Aronow BJ, Jegga AG. ToppGene Suite for gene list
enrichment analysis and candidate gene prioritization. Nucleic Acids Res.
2009;37(Web Server issue):W305-11. doi: 10.1093/nar/gkp427.

Himes, E. B, Jiang, X., Wagner, P., Hu, R., Wang, Q.,
Klanderman, B., Whitaker, M. R, Duan, Q., Lasky-Su, J., Nikolos,
C., Jester, W., Johnson, M., Panettieri, A. R, Tantisira, G. K, Weiss,
T. S, Lu, Q. (2014). “RNA-Seq Transcriptome Profiling Identifies CRISPLD2
as a Glucocorticoid Responsive Gene that Modulates Cytokine Function in
Airway Smooth Muscle Cells.” PLoS ONE, 9(6), e99625. http://www.ncbi.nlm.nih.gov/pubmed/24926665.

\url{https://www.bioconductor.org/packages/release/data/experiment/html/airway.html}
}
\usage{
data("toppdata.airway")
}
\description{
A dataframe of of sample toppData results created from the ifnb.de dataset using the toppFun() function
}
\keyword{datasets}
