% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toppdata.ifnb}
\alias{toppdata.ifnb}
\title{toppData example for ifnb.de}
\format{
A dataframe with 12,227 rows and 14 columns
\describe{
\item{Category}{ToppGene category}
\item{ID}{ToppGene Term ID}
\item{Name}{ToppGene Term Name}
\item{PValue}{P value}
\item{QValueFDRBH}{adjusted p-value (FDR)}
\item{QValueFDRBY}{adjusted p-value (BY)}
\item{QValueBonferroni}{adjusted p-value (Bonferroni)}
\item{TotalGenes}{Total genes in background}
\item{GenesInTerm}{Genes in ToppGene Term}
\item{GenesInQuery}{Genes in submitted query}
\item{GenesInTermQuery}{Intersection of genes in Term and in Query}
\item{Source}{ToppGene result source}
\item{URL}{ToppGene associated URL}
\item{Cluster}{cell group name}
}
}
\source{
Generated using ToppGene API (\url{https://toppgene.cchmc.org/}).
Chen J, Bardes EE, Aronow BJ, Jegga AG. ToppGene Suite for gene list
enrichment analysis and candidate gene prioritization. Nucleic Acids Res.
2009;37(Web Server issue):W305-11. doi: 10.1093/nar/gkp427.

\url{https://toppgene.cchmc.org}

Kang HM, Subramaniam M, Targ S, et al. Multiplexed droplet
single-cell RNA-sequencing using natural genetic variation.
Nat Biotechnol. 2018;36(1):89-94. doi:10.1038/nbt.4042
}
\usage{
data("toppdata.ifnb")
}
\description{
A dataframe of of sample toppData results created from the ifnb.de dataset using the toppFun() function
}
\keyword{datasets}
