% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toppdata.pbmc}
\alias{toppdata.pbmc}
\title{toppData example}
\format{
A dataframe with 8,550 rows and 14 columns
\describe{
\item{Category}{ToppGene category}
\item{ID}{ToppGene Term ID}
\item{Name}{ToppGene Term Name}
\item{PValue}{P value}
\item{QValueFDRBH}{adjusted p-value (FDR)}
\item{QValueFDRBY}{adjusted p-value (BY)}
\item{QValueBonferroni}{adjusted p-value (Bonferroni)}
\item{TotalGenes}{Total genes in background}
\item{GenesInTerm}{Genes in ToppGene Term}
\item{GenesInQuery}{Genes in submitted query}
\item{GenesInTermQuery}{Intersection of genes in Term and in Query}
\item{Source}{ToppGene result source}
\item{URL}{ToppGene associated URL}
\item{Cluster}{cell group name}
}
}
\source{
\url{https://toppgene.cchmc.org}

Generated using ToppGene API (\url{https://toppgene.cchmc.org/}).
Chen J, Bardes EE, Aronow BJ, Jegga AG. ToppGene Suite for gene list
enrichment analysis and candidate gene prioritization. Nucleic Acids Res.
2009;37(Web Server issue):W305-11. doi: 10.1093/nar/gkp427.

10X Genomics PBMC 3k dataset. Available from
\url{https://www.10xgenomics.com/resources/datasets/}.
Analysis following Seurat PBMC tutorial:
\url{https://satijalab.org/seurat/articles/pbmc3k_tutorial.html}
}
\usage{
data("toppdata.pbmc")
}
\description{
A dataframe of of sample toppData results created from the pbmc.markers dataset using the toppFun() function
}
\keyword{datasets}
