% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toppData}
\alias{toppData}
\title{toppData example}
\format{
A dataframe with 8,550 rows and 14 columns
\describe{
\item{Category}{ToppGene category}
\item{ID}{ToppGene Term ID}
\item{Name}{ToppGene Term Name}
\item{PValue}{P value}
\item{QValueFDRBH}{adjusted p-value (FDR)}
\item{QValueFDRBY}{adjusted p-value (BY)}
\item{QValueBonferroni}{adjusted p-value (Bonferroni)}
\item{TotalGenes}{Total genes in background}
\item{GenesInTerm}{Genes in ToppGene Term}
\item{GenesInQuery}{Genes in submitted query}
\item{GenesInTermQuery}{Intersection of genes in Term and in Query}
\item{Source}{ToppGene result source}
\item{URL}{ToppGene associated URL}
\item{Cluster}{cell group name}
}
}
\source{
\url{https://toppgene.cchmc.org}
}
\usage{
data("toppData")
}
\description{
A dataframe of of sample toppData results
}
\keyword{datasets}
