% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toppFun.R
\name{toppSave}
\alias{toppSave}
\title{Save toppData results (optionally) split by celltype/cluster}
\usage{
toppSave(
  toppData,
  filename = "toppData_results",
  save_dir = NULL,
  split = TRUE,
  format = "xlsx",
  cluster_col = "Cluster",
  verbose = TRUE
)
}
\arguments{
\item{toppData}{Results from toppFun as a dataframe}

\item{filename}{filename prefix for each split file}

\item{save_dir}{the directory to save files}

\item{split}{Boolean, whether to split the dataframe by celltype/cluster}

\item{format}{Saved file format, one of c("xlsx", "csv", "tsv")}

\item{cluster_col}{Column name for the groups of cells (e.g. cluster or celltype), usually "Cluster"}

\item{verbose}{Verbosity setting, TRUE or FALSE}
}
\value{
A saved file
}
\description{
Save toppData results (optionally) split by celltype/cluster
}
\examples{
data("toppdata.ifnb")
toppSave(toppdata.ifnb, 
    filename = "toppFun_results", 
    save_dir = tempdir(), 
    split = TRUE, 
    format = "xlsx")
}
