% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topp_plot.R
\name{toppBalloon}
\alias{toppBalloon}
\title{Create a balloon plot from toppdata results}
\usage{
toppBalloon(
  toppData,
  categories = NULL,
  balloons = 3,
  x_axis_text_size = 6,
  cluster_col = "Cluster",
  filename = "toppBalloon",
  save = FALSE,
  save_dir = NULL,
  height = 5,
  width = 10
)
}
\arguments{
\item{toppData}{A toppData results dataframe}

\item{categories}{The topp categories to plot}

\item{balloons}{Number of balloons per group to plot}

\item{x_axis_text_size}{Size of the text on the x axis}

\item{cluster_col}{The column name for clusters (default: "Cluster")}

\item{filename}{Filename of the saved balloon plot}

\item{save}{Save the balloon plot if TRUE}

\item{save_dir}{Directory to save the balloon plot}

\item{height}{Height of the saved balloon plot}

\item{width}{Width of the saved balloon plot}
}
\value{
ggplot
}
\description{
Create a balloon plot from toppdata results
}
\examples{
data("toppData")
toppBalloon(toppData, balloons = 3, save = FALSE)


}
