% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOPP_GENE.R
\name{toppFun}
\alias{toppFun}
\title{Get results from ToppFun}
\usage{
toppFun(
  markers,
  topp_categories,
  key_type = "SYMBOL",
  p_value = 0.05,
  min_genes = 1,
  max_genes = 1500,
  max_results = 10,
  correction = "FDR"
)
}
\arguments{
\item{markers}{A vector of markers or dataframe with columns as cluster labels}

\item{topp_categories}{A string or vector with specific toppfun categories for the query}

\item{key_type}{Gene name format}

\item{p_value}{P-value cutoff for results}

\item{min_genes}{Minimum number of genes to match in a query}

\item{max_genes}{Maximum number of genes to match in a query}

\item{max_results}{Maximum number of results per cluster}

\item{correction}{P-value correction method ("FDR" is "BH")}
}
\value{
data.frame
}
\description{
Get results from ToppFun
}
\examples{
toppFun(markers=c("IFNG", "FOXP3"), topp_categories="GeneOntologyBiologicalProcess", key_type="SYMBOL")
}
