% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOPP_GENE.R
\name{toppFun}
\alias{toppFun}
\title{Get results from ToppFun}
\usage{
toppFun(
  markers,
  topp_categories = NULL,
  cluster_col = "cluster",
  gene_col = "gene",
  p_val_col = "adj_p_val_col",
  logFC_col = "avg_logFC",
  num_genes = 1000,
  pval_cutoff = 0.5,
  fc_cutoff = 0,
  fc_filter = "ALL",
  clusters = NULL,
  correction = "FDR",
  key_type = "SYMBOL",
  min_genes = 2,
  max_genes = 1500,
  max_results = 50
)
}
\arguments{
\item{markers}{A vector of markers or dataframe with columns as cluster labels}

\item{topp_categories}{A string or vector with specific toppfun categories for the query}

\item{cluster_col}{Column name for the groups of cells (e.g. cluster or celltype)}

\item{gene_col}{Column name for genes (e.g. gene or feature)}

\item{p_val_col}{Column name for the p-value or adjusted p-value (preferred)}

\item{logFC_col}{Column name for the avg log FC column}

\item{num_genes}{Number of genes per group to use for toppGene query}

\item{pval_cutoff}{(adjusted) P-value cutoff for filtering differentially expressed genes}

\item{fc_cutoff}{Avg log fold change cutoff for filtering differentially expressed genes}

\item{fc_filter}{Include "ALL" genes, or only "UPREG" or "DOWNREG" for each cluster}

\item{clusters}{Which clusters to include in toppGene query}

\item{correction}{P-value correction method ("FDR" is "BH")}

\item{key_type}{Gene name format}

\item{min_genes}{Minimum number of genes to match in a query}

\item{max_genes}{Maximum number of genes to match in a query}

\item{max_results}{Maximum number of results per cluster}
}
\value{
data.frame
}
\description{
Get results from ToppFun
}
\examples{
data("ifnb.de")
toppData <- toppFun(ifnb.de,
    topp_categories = NULL,
    cluster_col = "celltype",
    gene_col = "gene",
    p_val_col = "p_val_adj",
    logFC_col = "avg_log2FC")
}
