% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toppFun.R
\name{toppFun}
\alias{toppFun}
\title{Get results from ToppFun}
\usage{
toppFun(
  input_data,
  type = "degs",
  topp_categories = NULL,
  cluster_col = "cluster",
  gene_col = "gene",
  p_val_col = "adj_p_val_col",
  logFC_col = "avg_logFC",
  direction_mode = "all",
  num_genes = 1000,
  pval_cutoff = 0.5,
  fc_cutoff = 0,
  fc_filter = "ALL",
  clusters = NULL,
  correction = "FDR",
  key_type = "SYMBOL",
  min_genes = 2,
  max_genes = 1500,
  max_results = 50,
  verbose = TRUE
)
}
\arguments{
\item{input_data}{A vector of markers or dataframe with columns as cluster labels}

\item{type}{One of c("degs", "marker_list", or "marker_df). If "degs" is selected,
the input_data is assumed to be a data.frame with logfoldchange, pvalue, and gene name columns.
If "marker_list" is selected, input_data is assumed to be a list of genes with no other stats,
and any thresholds pertaining to "degs" will be ignored.
If "marker_df" is selected, the input_data is assumed to be a data.frame with columns as clusters/celltypes,
and entries are lists of markers.}

\item{topp_categories}{A string or vector with specific toppfun categories for the query}

\item{cluster_col}{Column name for the groups of cells (e.g. cluster or celltype)}

\item{gene_col}{Column name for genes (e.g. gene or feature)}

\item{p_val_col}{Column name for the p-value or adjusted p-value (preferred)}

\item{logFC_col}{Column name for the avg log FC column}

\item{direction_mode}{One of c("all", "split"). Whether to use all genes in the pathway analysis, or to split by up and down regulated genes}

\item{num_genes}{Number of genes per group to use for toppGene query}

\item{pval_cutoff}{(adjusted) P-value cutoff for filtering differentially expressed genes}

\item{fc_cutoff}{Avg log fold change cutoff for filtering differentially expressed genes}

\item{fc_filter}{Include "ALL" genes, or only "UPREG" or "DOWNREG" for each cluster}

\item{clusters}{Which clusters to include in toppGene query}

\item{correction}{P-value correction method ("FDR" is "BH")}

\item{key_type}{Gene name format}

\item{min_genes}{Minimum number of genes to match in a query}

\item{max_genes}{Maximum number of genes to match in a query}

\item{max_results}{Maximum number of results per cluster}

\item{verbose}{Verbosity setting, TRUE or FALSE}
}
\value{
data.frame
}
\description{
The toppFun() function takes a data.frame or other tabular data structure and selects genes to use in querying
ToppGene.
}
\details{
The use of data from ToppGene is governed by their Terms of Use: https://toppgene.cchmc.org/navigation/termsofuse.jsp
}
\examples{
data("ifnb.de")
toppData <- toppFun(ifnb.de,
    topp_categories = NULL,
    cluster_col = "celltype",
    gene_col = "gene",
    p_val_col = "p_val_adj",
    logFC_col = "avg_log2FC"
)
}
