% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topp_plot.R
\name{toppPlot}
\alias{toppPlot}
\title{Create a dotplot from toppdata results
NEW: this accepts only 1 category}
\usage{
toppPlot(
  toppData,
  category,
  clusters = NULL,
  num_terms = 10,
  p_val_adj = "BH",
  p_val_display = "log",
  save = FALSE,
  save_dir = NULL,
  save_prefix = NULL,
  width = 5,
  height = 6,
  filename = NULL,
  y_axis_text_size = 8,
  combine = FALSE,
  ncols = NULL,
  ...
)
}
\arguments{
\item{toppData}{A toppData results dataframe}

\item{category}{The topp categories to plot}

\item{clusters}{The cluster(s) to plot}

\item{num_terms}{The number of terms from the toppData results to be plotted, per cluster}

\item{p_val_adj}{The P-value correction method: "BH", "Bonferroni", "BY", or "none"}

\item{save}{Whether to save the file automatically}

\item{save_dir}{Directory to save file}

\item{width}{width of the saved file (inches)}

\item{height}{height of the saved file (inches)}

\item{filename}{file name if saving the plot}

\item{y_axis_text_size}{Size of the Y axis text - for certain categories, it's helpful to decrease this}
}
\value{
ggplot
}
\description{
Create a dotplot from toppdata results
NEW: this accepts only 1 category
}
\examples{
toppPlot(toppData, category="GeneOntologyMolecularFunction", clusters=0, save=TRUE, save_prefix="MF_cluster0")

}
