% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topp_plot.R
\name{toppPlot}
\alias{toppPlot}
\title{Create a dotplot from toppdata results}
\usage{
toppPlot(
  toppData,
  category = NULL,
  clusters = NULL,
  num_terms = 10,
  p_val_adj = "BH",
  save = FALSE,
  save_dir = NULL,
  width = 7
)
}
\arguments{
\item{toppData}{A toppData results dataframe}

\item{category}{The topp category to plot}

\item{clusters}{The cluster(s) to plot}

\item{num_terms}{The number of terms from the toppData results to be plotted, per cluster}

\item{p_val_adj}{The P-value correction method: "BH", "Bonferroni", "BY", or "none"}
}
\value{
ggplot
}
\description{
Create a dotplot from toppdata results
}
