% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toppFun.R
\name{toppSave}
\alias{toppSave}
\title{Save toppData results (optionally) split by celltype/cluster}
\usage{
toppSave(
  toppData,
  filename = NULL,
  save_dir = NULL,
  split = TRUE,
  format = "xlsx"
)
}
\arguments{
\item{toppData}{Results from toppFun as a dataframe}

\item{filename}{filename prefix for each split file}

\item{save_dir}{the directory to save files}

\item{split}{Boolean, whether to split the dataframe by celltype/cluster}

\item{format}{Saved file format, one of c("xlsx", "csv", "tsv")}
}
\value{
A saved file
}
\description{
Save toppData results (optionally) split by celltype/cluster
}
\examples{
data("toppData")
toppSave(toppData, filename="toppFun_results", split = TRUE, format = "xlsx")
}
