% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOPP_GENE.R
\name{toppSave}
\alias{toppSave}
\title{Save toppData results (optionally) split by celltype/cluster}
\usage{
toppSave(toppData, filename = NULL, split = TRUE, format = "xlsx")
}
\arguments{
\item{toppData}{Results from toppFun as a dataframe}

\item{filename}{filename prefix for each split file}

\item{split}{Boolean, whether to split the dataframe by celltype/cluster}

\item{format}{Saved file format, one of c("xlsx", "csv", "tsv")}
}
\value{
A saved file
}
\description{
Save toppData results (optionally) split by celltype/cluster
}
\examples{
toppSave(toppData, filename="toppFun_results", split = TRUE, format = "xlsx")
}
